mcom03rtl_defconfig: Add rtl-test-stop package
Provides a stop marker for a RTL test.
The test will be successfuly completed if the RTL
test observes a magic value in a magic RAM address.

A standalone app is less than Bash shell, thus allows
to reduce a initramfs size and a RTL test duration.

The package creates a tiny initramfs and replaces
an initramfs which is generated by the Buildroot.

Issue: #MCOM03SW-156
Change-Id: I73f62c770691c47dc62df466900ff45ae4a9cf1a