defconfigs: Add new common defconfig and fragments
The root filesystem configuration is interchangeable between all
boards, but the SBL (and its components) are not. Therefore, it
makes sense to reduce number of defconfigs and store root filesystem
and SBL configurations separately.

Configuration for each board is composed of defconfig for root
filesystem (mcom03_defconfig), and one or several fragment,
for example mcom03bub.fragment.

Old defconfigs will be removed when all repositories and CI switch
to new defconfig.

Issue: #MCOM03SW-1296
Change-Id: Idda74d0c923b69466a98d739027140a3264d8f9a