package/mcom03-felix: Specify Linux version
It is required to know Linux version during exporting or importing
binary distribution. The variable LINUX_VERSION_PROBED cannot be used
as it uses a shell command that can fail if Linux is not built yet.
As well the variable BR2_LINUX_KERNEL_CUSTOM_REPO_VERSION cannot be
used as it does not include Linux version in general.

So the variable BR2_PACKAGE_MCOM03_FELIX_LINUX_VERSION was
defined and used.

Issue: #MCOM03SW-1922
Change-Id: I23b4ab96f2c25a09dce44e005dc48d4b36813ca3