package/mcom03-rogue-libs: Create tar with staging files
mcom03-rogue-libs packages provides OpenCL and GLES headers, but we
want to have ability to build OpenCL and GLES applications without
building mcom03-rogue-libs itself. To achieve this we need to create
archives with target and staging files during the private build and
use them when build is public.

Issue: #MCOM03SW-2573
Depends-On: https://gerrit.elvees.com/c/buildroot-external/common/+/43291
Depends-On: https://gerrit.elvees.com/c/tools/buildroot/+/43292
Change-Id: I0b4583551d0f75e5b14d8e5300130d348a45cfc0