package/mcom03-felix: Show all options always
The package recipe enables to build mcom03-felix from source code, if
the package is present in local.mk file. The feature was added in the
I26f5e33e40e.

But if just the MCOM03_FELIX_OVERRIDE_SRC was enabled, then other
options like BR2_PACKAGE_MCOM03_FELIX_SENSOR_* were disabled anyway. It
means, that the build was partial, because sensor configs, utilities
were not installed to the target FS.

To resolve the situation we display all options always. The cases:

* INSTALL_SRC isn't set, OVERRIDE_SRCDIR is present: the usual build
  from source from the path in the OVERRIDE_SRCDIR;
* INSTALL_SRC isn't set, OVERRIDE_SRCDIR isn't present: just use
  prebuilded archive;
* INSTALL_SRC is set, OVERRIDE_SRCDIR isn't present: the usual build
  from source;
* INSTALL_SRC is set, OVERRIDE_SRCDIR is present: the usual build from
  source from the path in the OVERRIDE_SRCDIR.

Issue: #MCOM03SW-3529
Change-Id: I0843d76781c8fd0a11fd26c5cc9d96b30f7ece79