overlay: mcom03fc: Disable systemd-pstore
systemd-pstore.service is a systemd service that archives the contents
of the Linux persistent storage filesystem, pstore, to other storage,
thus preserving the existing information contained in the pstore,
and clearing pstore storage for future error events.

It is not required, so disabling makes boot time to be less.

Issue: #MCOM03SW-3582
Change-Id: Ibc2540764faf48bca0ad85b91c97e75239047198