overlay: mcom03fc: Disable systemd-timesyncd
systemd-timesyncd.service is a systemd service that may be used to
synchronize the local system clock with a remote Network Time Protocol
(NTP) server. It also saves the local time to disk every time the clock
has been synchronized and uses this to possibly advance the system
realtime clock on subsequent reboots to ensure it (roughly)
monotonically advances even if the system lacks a battery-buffered RTC
chip.

It is not required, so disabling makes boot time to be less.

Issue: #MCOM03SW-3582
Change-Id: Ib6565724f1cf922613fa6cf91ea9681661cf2aa2