mcom03fc_defconfig: Mask systemd-fsck-root service
The systemd-fsck-root.service is responsible for file system checks on
the root file system, but only if the root filesystem was not checked
in the initramfs.

This service is started at boot if passno in /etc/fstab for the file
system is set to a value greater than zero. The file system check for
root is performed before the other file systems. Other file systems may
be checked in parallel, except when they are on the same rotating disk.

It is not required, so masking makes boot time to be less.

Issue: #MCOM03SW-3582
Change-Id: I275c7d6e9455cc7ff157183d6ac343fa2e50c022