mcom03fc_defconfig: Add post fakeroot script
To fully disable required services it is required to use the systemd
mask operation but the systemd preset is not allowed to do the such
operation. Let's replace the systemd presets added in the following
changes with the script running as a Buildroot post fakeroot hook:

* Ib6565724f1cf922
* I7966d94174e5bd1
* Ibc2540764faf48b
* I113f9c70af9bc01

According to Buildroot documentation the script is running after all
finalization hooks to be done (including applying overlays and systemd
preset) to ensure it can override any late configuration done by
packages. The script is called with the target directory name as first
argument. The rest arguments are passed via BR2_ROOTFS_POST_SCRIPT_ARGS
variable.

The following optional arguments currently are supported:

* -m   comma-separated systemd units to mask;
* -a   colon-separated list of files or directories additionally
       to be appended to target rootfs.

Issue: #MCOM03SW-3611
Issue: #MCOM03SW-3582
Depends-On: https://gerrit.elvees.com/c/mcom03/buildroot/+/53295
Change-Id: I08483fc5ed847ac5a9f5f2cd851cfef800e168f1