configs: Enable spidev_test package
The utility will be used for SPI testing.

The package is enabled for mcom03r-fc_defconfig and mcom03r_defconfig
only because we plan to use it at the IP_KU board only for now.

Formally we need the utility in mcom03r-fc_defconfig only, because
consumers don't have an IP_KU board; hence they will not be able to
reproduce our tests. But they may want to use the utility for other
purposes, so let's enable it for mcom03r_defconfig too.

Issue: #MCOM03SW-3528
Change-Id: I80979221bddd5da232eb55b78ab16caf67211f9e